Register exporters in main
